\name{simesants}
\alias{simesants}
\docType{data}
\title{Harvard Forest Simes Tract Ant Community Data}
\description{
Within the Simes tract of Harvard Forest, we sampled ants by hand collection along 100-m transects at eight sites with the following microhabitats: two sites each of hemlock, white pine, and hardwood overstories; a transition between hardwood swamp and pine; and a rocky-slope with a hardwood overstory.
}
\usage{data(simesants)}
\format{
  A data frame with 5 observations on the following 23 variables.
  \describe{
    \item{\code{Habitat}}{a factor with levels \code{Hardwood} \code{Hardwood_Pine_Swamp} \code{Hardwood_Rocky_Slope} \code{Hemlock} \code{White_Pine}}
    \item{\code{Acasub}}{a numeric vector}
    \item{\code{Aphrud}}{a numeric vector}
    \item{\code{Camher}}{a numeric vector}
    \item{\code{Campen}}{a numeric vector}
    \item{\code{Crecer}}{a numeric vector}
    \item{\code{Crelin}}{a numeric vector}
    \item{\code{Dolpla}}{a numeric vector}
    \item{\code{Dolpus}}{a numeric vector}
    \item{\code{Forneo}}{a numeric vector}
    \item{\code{Forsuba}}{a numeric vector}
    \item{\code{Lasali}}{a numeric vector}
    \item{\code{Lasnea}}{a numeric vector}
    \item{\code{Lasspe}}{a numeric vector}
    \item{\code{Lasumb}}{a numeric vector}
    \item{\code{Myrpun}}{a numeric vector}
    \item{\code{Myrscu}}{a numeric vector}
    \item{\code{Myrsmi}}{a numeric vector}
    \item{\code{Stebre}}{a numeric vector}
    \item{\code{Steimp}}{a numeric vector}
    \item{\code{Stesch}}{a numeric vector}
    \item{\code{Tapses}}{a numeric vector}
    \item{\code{Temlon}}{a numeric vector}
  }
}
\details{
  Columns correspond to the ant species, designated by a six letter code combing the first three letters of the genus with the first three letters of the species.   
}
\source{
\url{http://harvardforest.fas.harvard.edu/data/archive.html}
}

\examples{
data(simesants)
}
\keyword{datasets}
